+++ contracts/osmo-liquid-pooler/src/msg.rs
@@ -6,7 +6,7 @@ use cosmwasm_std::{
     WasmMsg,
 };
 use covenant_macros::{clocked, covenant_clock_address, covenant_deposit_address};
-use covenant_utils::OutpostExecuteMsg;
+use covenant_utils::osmo::OutpostExecuteMsg;
 use polytone::callbacks::CallbackMessage;
 
 #[cw_serde]
@@ -91,7 +91,7 @@ impl LiquidityProvisionConfig {
             funds.push(c.clone());
         }
 
-        let outpost_config = covenant_utils::OutpostProvideLiquidityConfig {
+        let outpost_config = covenant_utils::osmo::OutpostProvideLiquidityConfig {
             pool_id: Uint64::new(self.pool_id.u64()),
             expected_spot_price: self.expected_spot_price,
             acceptable_price_spread: self.acceptable_price_spread,
