+++ contracts/two-party-pol-holder/src/msg.rs
@@ -10,7 +10,7 @@ use covenant_clock::helpers::dequeue_msg;
 use covenant_macros::{
     clocked, covenant_clock_address, covenant_deposit_address, covenant_next_contract,
 };
-use covenant_utils::{DenomSplit, SplitConfig, SplitType};
+use covenant_utils::split::{DenomSplit, SplitConfig, SplitType};
 use cw_utils::Expiration;
 
 use crate::{error::ContractError, state::CONTRACT_STATE};
