+++ packages/covenant-utils/src/lib.rs
@@ -1,407 +1,19 @@
+pub mod astroport;
 pub mod deadline;
 pub mod instantiate2_helper;
+pub mod neutron_ica;
+pub mod osmo;
+pub mod polytone;
+pub mod split;
 pub mod withdraw_lp_helper;
 
-use astroport::asset::PairInfo;
-use cosmwasm_schema::{cw_serde, QueryResponses};
+use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{
-    to_json_binary, Addr, Attribute, BankMsg, Binary, BlockInfo, Coin, CosmosMsg, Decimal, Empty,
-    Fraction, IbcMsg, IbcTimeout, QuerierWrapper, QueryRequest, StdError, StdResult, Timestamp,
+    Addr, Attribute, BankMsg, BlockInfo, Coin, CosmosMsg, IbcMsg, IbcTimeout, StdError, Timestamp,
     Uint128, Uint64,
 };
-use cw20::BalanceResponse;
-use neutron_sdk::{
-    bindings::msg::{IbcFee, NeutronMsg},
-    sudo::msg::RequestPacketTimeoutHeight,
-};
-use polytone::callbacks::CallbackRequest;
-use std::collections::BTreeMap;
-
-pub mod neutron_ica {
-    use cosmwasm_schema::{cw_serde, QueryResponses};
-    use cosmwasm_std::{Attribute, Binary, Coin, StdError, Uint128, Uint64};
-    use neutron_sdk::{
-        bindings::{msg::IbcFee, types::ProtobufAny},
-        NeutronResult,
-    };
-    use prost::Message;
-
-    #[cw_serde]
-    pub struct OpenAckVersion {
-        pub version: String,
-        pub controller_connection_id: String,
-        pub host_connection_id: String,
-        pub address: String,
-        pub encoding: String,
-        pub tx_type: String,
-    }
-
-    /// SudoPayload is a type that stores information about a transaction that we try to execute
-    /// on the host chain. This is a type introduced for our convenience.
-    #[cw_serde]
-    pub struct SudoPayload {
-        pub message: String,
-        pub port_id: String,
-    }
-
-    /// Serves for storing acknowledgement calls for interchain transactions
-    #[cw_serde]
-    pub enum AcknowledgementResult {
-        /// Success - Got success acknowledgement in sudo with array of message item types in it
-        Success(Vec<String>),
-        /// Error - Got error acknowledgement in sudo with payload message in it and error details
-        Error((String, String)),
-        /// Timeout - Got timeout acknowledgement in sudo with payload message in it
-        Timeout(String),
-    }
-
-    #[cw_serde]
-    pub struct RemoteChainInfo {
-        /// connection id from neutron to the remote chain on which
-        /// we wish to open an ICA
-        pub connection_id: String,
-        pub channel_id: String,
-        pub denom: String,
-        pub ibc_transfer_timeout: Uint64,
-        pub ica_timeout: Uint64,
-        pub ibc_fee: IbcFee,
-    }
-
-    impl RemoteChainInfo {
-        pub fn get_response_attributes(&self) -> Vec<Attribute> {
-            let recv_fee = coin_vec_to_string(&self.ibc_fee.recv_fee);
-            let ack_fee = coin_vec_to_string(&self.ibc_fee.ack_fee);
-            let timeout_fee = coin_vec_to_string(&self.ibc_fee.timeout_fee);
-
-            vec![
-                Attribute::new("connection_id", &self.connection_id),
-                Attribute::new("channel_id", &self.channel_id),
-                Attribute::new("denom", &self.denom),
-                Attribute::new(
-                    "ibc_transfer_timeout",
-                    self.ibc_transfer_timeout.to_string(),
-                ),
-                Attribute::new("ica_timeout", self.ica_timeout.to_string()),
-                Attribute::new("ibc_recv_fee", recv_fee),
-                Attribute::new("ibc_ack_fee", ack_fee),
-                Attribute::new("ibc_timeout_fee", timeout_fee),
-            ]
-        }
-
-        pub fn validate(self) -> Result<RemoteChainInfo, StdError> {
-            if self.ibc_fee.ack_fee.is_empty()
-                || self.ibc_fee.timeout_fee.is_empty()
-                || !self.ibc_fee.recv_fee.is_empty()
-            {
-                return Err(StdError::generic_err("invalid IbcFee".to_string()));
-            }
-
-            Ok(self)
-        }
-    }
-
-    fn coin_vec_to_string(coins: &Vec<Coin>) -> String {
-        let mut str = "".to_string();
-        if coins.is_empty() {
-            str.push_str("[]");
-        } else {
-            for coin in coins {
-                str.push_str(&coin.to_string());
-            }
-        }
-        str.to_string()
-    }
-
-    pub fn get_proto_coin(
-        denom: String,
-        amount: Uint128,
-    ) -> cosmos_sdk_proto::cosmos::base::v1beta1::Coin {
-        cosmos_sdk_proto::cosmos::base::v1beta1::Coin {
-            denom,
-            amount: amount.to_string(),
-        }
-    }
-
-    #[cw_serde]
-    #[derive(QueryResponses)]
-    pub enum QueryMsg {
-        /// Returns the associated remote chain information
-        #[returns(Option<String>)]
-        DepositAddress {},
-    }
-
-    #[cw_serde]
-    #[derive(QueryResponses)]
-    pub enum CovenantQueryMsg {
-        /// Returns the associated remote chain information
-        #[returns(Option<String>)]
-        DepositAddress {},
-    }
-
-    /// helper that serializes a MsgTransfer to protobuf
-    pub fn to_proto_msg_transfer(msg: impl Message) -> NeutronResult<ProtobufAny> {
-        // Serialize the Transfer message
-        let mut buf = Vec::with_capacity(msg.encoded_len());
-        if let Err(e) = msg.encode(&mut buf) {
-            return Err(StdError::generic_err(format!("Encode error: {e}")).into());
-        }
-
-        Ok(ProtobufAny {
-            type_url: "/ibc.applications.transfer.v1.MsgTransfer".to_string(),
-            value: Binary::from(buf),
-        })
-    }
-
-    pub fn to_proto_msg_send(msg: impl Message) -> NeutronResult<ProtobufAny> {
-        // Serialize the Send message
-        let mut buf = Vec::with_capacity(msg.encoded_len());
-        if let Err(e) = msg.encode(&mut buf) {
-            return Err(StdError::generic_err(format!("Encode error: {e}")).into());
-        }
-
-        Ok(ProtobufAny {
-            type_url: "/cosmos.bank.v1beta1.MsgSend".to_string(),
-            value: Binary::from(buf),
-        })
-    }
-
-    pub fn to_proto_msg_multi_send(msg: impl Message) -> NeutronResult<ProtobufAny> {
-        // Serialize the Send message
-        let mut buf = Vec::with_capacity(msg.encoded_len());
-        if let Err(e) = msg.encode(&mut buf) {
-            return Err(StdError::generic_err(format!("Encode error: {e}")).into());
-        }
-
-        Ok(ProtobufAny {
-            type_url: "/cosmos.bank.v1beta1.MsgMultiSend".to_string(),
-            value: Binary::from(buf),
-        })
-    }
-}
-
-// splitter
-#[cw_serde]
-pub struct DenomSplit {
-    pub denom: String,
-    pub split: SplitType,
-}
-
-#[cw_serde]
-pub enum SplitType {
-    Custom(SplitConfig),
-    // predefined splits will go here
-}
-
-impl SplitType {
-    pub fn get_split_config(self) -> Result<SplitConfig, StdError> {
-        match self {
-            SplitType::Custom(c) => Ok(c),
-        }
-    }
-}
-
-#[cw_serde]
-pub struct SplitConfig {
-    /// map receiver address to its share of the split
-    pub receivers: BTreeMap<String, Decimal>,
-}
-
-impl SplitConfig {
-    pub fn remap_receivers_to_routers(
-        &self,
-        receiver_a: String,
-        router_a: String,
-        receiver_b: String,
-        router_b: String,
-    ) -> Result<SplitConfig, StdError> {
-        let mut new_receivers = BTreeMap::new();
-
-        match self.receivers.get(&receiver_a) {
-            Some(val) => new_receivers.insert(router_a, *val),
-            None => {
-                return Err(StdError::NotFound {
-                    kind: format!("receiver {receiver_b:?} not found"),
-                })
-            }
-        };
-        match self.receivers.get(&receiver_b) {
-            Some(val) => new_receivers.insert(router_b, *val),
-            None => {
-                return Err(StdError::NotFound {
-                    kind: format!("receiver {receiver_b:?} not found"),
-                })
-            }
-        };
-
-        Ok(SplitConfig {
-            receivers: new_receivers,
-        })
-    }
-
-    pub fn validate(&self, party_a: &str, party_b: &str) -> Result<(), StdError> {
-        let share_a = match self.receivers.get(party_a) {
-            Some(val) => *val,
-            None => return Err(StdError::not_found(party_a)),
-        };
-        let share_b = match self.receivers.get(party_b) {
-            Some(val) => *val,
-            None => return Err(StdError::not_found(party_b)),
-        };
-
-        if share_a + share_b != Decimal::one() {
-            return Err(StdError::generic_err(
-                "shares must add up to 1.0".to_string(),
-            ));
-        }
-
-        Ok(())
-    }
-
-    /// Validate that all shares are added to one
-    pub fn validate_shares(&self) -> Result<(), StdError> {
-        let mut total_shares = Decimal::zero();
-
-        for (_, share) in self.receivers.clone() {
-            total_shares += share;
-        }
-
-        if total_shares != Decimal::one() {
-            return Err(StdError::generic_err(
-                "shares must add up to 1.0".to_string(),
-            ));
-        }
-
-        Ok(())
-    }
-
-    pub fn get_transfer_messages(
-        &self,
-        amount: Uint128,
-        denom: String,
-        filter_addr: Option<String>,
-    ) -> Result<Vec<CosmosMsg>, StdError> {
-        let msgs: Result<Vec<CosmosMsg>, StdError> = self
-            .receivers
-            .iter()
-            .map(|(addr, share)| {
-                // if we are filtering for a single receiver,
-                // then we wish to transfer only to that receiver.
-                // we thus set receiver share to 1.0, as the
-                // entitlement already takes that into account.
-                match &filter_addr {
-                    Some(filter) => {
-                        if filter == addr {
-                            (addr, Decimal::one())
-                        } else {
-                            (addr, Decimal::zero())
-                        }
-                    }
-                    None => (addr, *share),
-                }
-            })
-            .filter(|(_, share)| !share.is_zero())
-            .map(|(addr, share)| {
-                let entitlement = amount
-                    .checked_multiply_ratio(share.numerator(), share.denominator())
-                    .map_err(|_| StdError::generic_err("failed to checked_multiply".to_string()))?;
-
-                let amount = Coin {
-                    denom: denom.to_string(),
-                    amount: entitlement,
-                };
-
-                Ok(CosmosMsg::Bank(BankMsg::Send {
-                    to_address: addr.to_string(),
-                    amount: vec![amount],
-                }))
-            })
-            .collect();
-
-        msgs
-    }
-
-    pub fn get_response_attribute(&self, denom: String) -> Attribute {
-        let mut receivers = "[".to_string();
-        self.receivers.iter().for_each(|(receiver, share)| {
-            receivers.push('(');
-            receivers.push_str(receiver);
-            receivers.push(':');
-            receivers.push_str(&share.to_string());
-            receivers.push_str("),");
-        });
-        receivers.push(']');
-        Attribute::new(denom, receivers)
-    }
-}
-
-// /// enum based configuration for asserting expiration.
-// /// works by asserting the current block against enum variants.
-// #[cw_serde]
-// pub enum ExpiryConfig {
-//     /// no expiration configured
-//     None,
-//     /// block height based expiry config
-//     Block(u64),
-//     /// timestamp based expiry config
-//     Time(Timestamp),
-// }
-
-// impl ExpiryConfig {
-//     pub fn get_response_attributes(&self) -> Vec<Attribute> {
-//         match self {
-//             ExpiryConfig::None => vec![Attribute::new("expiry_config", "none")],
-//             ExpiryConfig::Block(h) => vec![Attribute::new(
-//                 "expiry_config_expiry_block_height",
-//                 h.to_string(),
-//             )],
-//             ExpiryConfig::Time(t) => vec![Attribute::new(
-//                 "expiry_config_expiry_block_timestamp",
-//                 t.to_string(),
-//             )],
-//         }
-//     }
-
-//     /// validates that the lockup config being stored is not already expired.
-//     pub fn validate(&self, block_info: &BlockInfo) -> Result<(), StdError> {
-//         match self {
-//             ExpiryConfig::None => Ok(()),
-//             ExpiryConfig::Block(h) => {
-//                 if h > &block_info.height {
-//                     Ok(())
-//                 } else {
-//                     Err(StdError::generic_err(
-//                         "invalid expiry config: block height must be in the future".to_string(),
-//                     ))
-//                 }
-//             }
-//             ExpiryConfig::Time(t) => {
-//                 if t.nanos() > block_info.time.nanos() {
-//                     Ok(())
-//                 } else {
-//                     Err(StdError::generic_err(
-//                         "invalid expiry config: block time must be in the future".to_string(),
-//                     ))
-//                 }
-//             }
-//         }
-//     }
-
-//     /// compares current block info with the stored expiry config.
-//     /// returns false if no expiry configuration is stored.
-//     /// otherwise, returns true if the current block is past the stored info.
-//     pub fn is_expired(&self, block_info: BlockInfo) -> bool {
-//         match self {
-//             // no expiration date
-//             ExpiryConfig::None => false,
-//             // if stored expiration block height is less than or equal to the current block,
-//             // expired
-//             ExpiryConfig::Block(h) => h <= &block_info.height,
-//             // if stored expiration timestamp is more than or equal to the current timestamp,
-//             // expired
-//             ExpiryConfig::Time(t) => t.nanos() <= block_info.time.nanos(),
-//         }
-//     }
-// }
+use neutron_ica::{default_ibc_fee, get_default_ibc_fee_requirement};
+use neutron_sdk::{bindings::msg::NeutronMsg, sudo::msg::RequestPacketTimeoutHeight};
 
 #[cw_serde]
 pub enum ReceiverConfig {
@@ -548,40 +160,6 @@ pub struct DestinationConfig {
     pub ibc_transfer_timeout: Uint64,
 }
 
-pub fn default_ibc_ack_fee_amount() -> Uint128 {
-    Uint128::new(100000)
-}
-
-pub fn default_ibc_timeout_fee_amount() -> Uint128 {
-    Uint128::new(100000)
-}
-
-pub fn default_ibc_fee() -> IbcFee {
-    IbcFee {
-        // must be empty
-        recv_fee: vec![],
-        ack_fee: vec![cosmwasm_std::Coin {
-            denom: "untrn".to_string(),
-            amount: default_ibc_ack_fee_amount(),
-        }],
-        timeout_fee: vec![cosmwasm_std::Coin {
-            denom: "untrn".to_string(),
-            amount: default_ibc_timeout_fee_amount(),
-        }],
-    }
-}
-
-pub fn get_default_ibc_fee_requirement() -> Uint128 {
-    default_ibc_ack_fee_amount() + default_ibc_timeout_fee_amount()
-}
-
-pub fn get_default_ica_fee() -> Coin {
-    Coin {
-        denom: "untrn".to_string(),
-        amount: Uint128::new(1000000),
-    }
-}
-
 impl DestinationConfig {
     pub fn get_ibc_transfer_messages_for_coins(
         &self,
@@ -648,135 +226,4 @@ impl DestinationConfig {
             Attribute::new("ibc_transfer_timeout", self.ibc_transfer_timeout),
         ]
     }
-}
-
-/// queries the liquidity token balance of given address
-pub fn query_liquidity_token_balance(
-    querier: QuerierWrapper,
-    liquidity_token: &str,
-    contract_addr: String,
-) -> Result<Uint128, StdError> {
-    let liquidity_token_balance: BalanceResponse = querier.query_wasm_smart(
-        liquidity_token,
-        &cw20::Cw20QueryMsg::Balance {
-            address: contract_addr,
-        },
-    )?;
-    Ok(liquidity_token_balance.balance)
-}
-
-/// queries the cw20 liquidity token address corresponding to a given pool
-pub fn query_liquidity_token_address(
-    querier: QuerierWrapper,
-    pool: String,
-) -> Result<String, StdError> {
-    let pair_info: PairInfo =
-        querier.query_wasm_smart(pool, &astroport::pair::QueryMsg::Pair {})?;
-    Ok(pair_info.liquidity_token.to_string())
-}
-
-pub fn query_astro_pool_token(
-    querier: QuerierWrapper,
-    pool: String,
-    addr: String,
-) -> Result<AstroportPoolTokenResponse, StdError> {
-    let pair_info: PairInfo =
-        querier.query_wasm_smart(pool, &astroport::pair::QueryMsg::Pair {})?;
-
-    let liquidity_token_balance: BalanceResponse = querier.query_wasm_smart(
-        pair_info.liquidity_token.as_ref(),
-        &cw20::Cw20QueryMsg::Balance { address: addr },
-    )?;
-
-    Ok(AstroportPoolTokenResponse {
-        pair_info,
-        balance_response: liquidity_token_balance,
-    })
-}
-
-#[cw_serde]
-pub struct AstroportPoolTokenResponse {
-    pub pair_info: PairInfo,
-    pub balance_response: BalanceResponse,
-}
-
-#[cw_serde]
-pub enum PolytoneExecuteMsg {
-    Query {
-        msgs: Vec<QueryRequest<Empty>>,
-        callback: CallbackRequest,
-        timeout_seconds: Uint64,
-    },
-    Execute {
-        msgs: Vec<CosmosMsg<Empty>>,
-        callback: Option<CallbackRequest>,
-        timeout_seconds: Uint64,
-    },
-}
-
-#[cw_serde]
-#[derive(QueryResponses)]
-pub enum PolytoneQueryMsg {
-    #[returns(Option<String>)]
-    RemoteAddress { local_address: String },
-    #[returns(Uint64)]
-    BlockMaxGas,
-}
-
-pub fn get_polytone_execute_msg_binary(
-    msgs: Vec<CosmosMsg>,
-    callback: Option<CallbackRequest>,
-    timeout_seconds: Uint64,
-) -> StdResult<Binary> {
-    let execute_msg = PolytoneExecuteMsg::Execute {
-        msgs,
-        callback,
-        timeout_seconds,
-    };
-    to_json_binary(&execute_msg)
-}
-
-pub fn get_polytone_query_msg_binary(
-    msgs: Vec<QueryRequest<Empty>>,
-    callback: CallbackRequest,
-    timeout_seconds: Uint64,
-) -> StdResult<Binary> {
-    let query_msg = PolytoneExecuteMsg::Query {
-        msgs,
-        callback,
-        timeout_seconds,
-    };
-    to_json_binary(&query_msg)
-}
-
-pub fn query_polytone_proxy_address(
-    local_address: String,
-    note_address: String,
-    querier: QuerierWrapper,
-) -> Result<Option<String>, StdError> {
-    let remote_address_query = PolytoneQueryMsg::RemoteAddress { local_address };
-
-    querier.query_wasm_smart(note_address, &remote_address_query)
-}
-
-#[cw_serde]
-pub struct OutpostProvideLiquidityConfig {
-    /// id of the pool we wish to provide liquidity to
-    pub pool_id: Uint64,
-    /// the price which we expect to provide liquidity at
-    pub expected_spot_price: Decimal,
-    /// acceptable delta (both ways) of the expected price
-    pub acceptable_price_spread: Decimal,
-    /// slippage tolerance
-    pub slippage_tolerance: Decimal,
-    /// limits for single-side liquidity provision
-    pub asset_1_single_side_lp_limit: Uint128,
-    pub asset_2_single_side_lp_limit: Uint128,
-}
-
-#[cw_serde]
-pub enum OutpostExecuteMsg {
-    ProvideLiquidity {
-        config: OutpostProvideLiquidityConfig,
-    },
 }
